+++ src/WDCMethods.js
@@ -36,6 +36,8 @@ const formatJSONAsTable = (data, tableName) => {
   paramIndices.forEach(i => {
     let newEntry = {
       dateTime: tableSeries.values[0].value[i].dateTime,
+      latitude: tableSeries.sourceInfo.geoLocation.geogLocation.latitude,
+      longitude: tableSeries.sourceInfo.geoLocation.geogLocation.latitude,
       [tableName]: tableSeries.values[0].value[i].value
     };
     tableData.push(newEntry);
@@ -48,7 +50,6 @@ const formatJSONAsTable = (data, tableName) => {
 generates a URL for query parameters contained in the connectionData object accepted as an argument
 */
 const generateURL = connectionData => {
-  //todo standardize this template's format when we add more query info fields
   let paramList = connectionData.paramNums.replace(/\s/g, "").split(","); // split by comma, ignoring whitespace
   let paramQuery = `&parameterCd=${paramList.join()}`;
 
@@ -97,9 +98,18 @@ const generateSchemaTablesFromData = data => {
     cols.push({
       id: "dateTime",
       alias: "dateTime",
-      dataType: tableau.dataTypeEnum.string //placeholder until we develop connectionData more
+      dataType: tableau.dataTypeEnum.string
     });
-
+    cols.push({
+      id: "latitude",
+      alias: "latitude",
+      dataType: tableau.dataTypeEnum.float
+    });
+    cols.push({
+      id: "longitude",
+      alias: "longitude",
+      dataType: tableau.dataTypeEnum.float
+    });
     let name = series.name;
     let nameTokens = name.split(":");
     let site = nameTokens[1];
@@ -108,11 +118,11 @@ const generateSchemaTablesFromData = data => {
     cols.push({
       id: column,
       alias: column,
-      dataType: tableau.dataTypeEnum.string //placeholder until we develop connectionData more
+      dataType: tableau.dataTypeEnum.string
     });
     let newSchema = {
       id: column,
-      alias: "useful information will be put here", //todo, add useful information
+      alias: column,
       columns: cols
     };
     tableList.push(newSchema);