+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - inactive location parameter forms are now hidden
 - added input form for hydrological unit codes
 - added input form for up to 10 counties
+- added longitude and latitude columns to data Schema
 ### Changed